% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taigr.R
\name{download.raw.from.taiga}
\alias{download.raw.from.taiga}
\title{Download a "raw" file to cache directory.}
\usage{
download.raw.from.taiga(data.id = NULL, data.name = NULL,
  data.version = NULL, data.dir = "~/.taiga", force.taiga = FALSE,
  taiga.url = getOption("default.taiga.url", "https://cds.team/taiga"),
  quiet = FALSE, data.file = NULL)
}
\value{
the file path to the downloaded file
}
\description{
Download a "raw" file to cache directory.
}
