% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taigr.R
\name{get_agnostic_data}
\alias{get_agnostic_data}
\title{Get all type of data and set them by `filename` in the hash passed in parameter
`get_agnostic_data` doesn't return anything, the hash is mutated}
\usage{
get_agnostic_data(data.file, datasetVersion.id, dataset.name,
  dataset.version, taiga.url, force = FALSE, token = token,
  dict_filenames_data)
}
\arguments{
\item{data.file}{Datafile object received from taiga (not the data)}

\item{datasetVersion.id}{Id of the datasetVersion}

\item{dataset.name}{Name of the dataset}

\item{dataset.version}{Version number of the datasetVersion}

\item{taiga.url}{Url to contact Taiga}

\item{force}{Boolean to ask Taiga if it needs to do the conversion again (in case of failure or interruption)}

\item{token}{Path of the location of the token file containing the Taiga token of the user}

\item{dict_filenames_data}{Hash which is going to contain the data in the form `filename`: `data`}
}
\value{
This function does not return anything, but mutate a hash object
}
\description{
Get all type of data and set them by `filename` in the hash passed in parameter
`get_agnostic_data` doesn't return anything, the hash is mutated
}
