% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taigr.R
\name{load.all.datafiles.from.taiga}
\alias{load.all.datafiles.from.taiga}
\title{Function to retrieve all the data available in a specific datasetVersion}
\usage{
load.all.datafiles.from.taiga(datasetVersion.id = NULL,
  dataset.name = NULL, dataset.version = NULL, transpose = FALSE,
  cache.dir = "~/.taiga", force.taiga = FALSE,
  taiga.url = getOption("default.taiga.url", "https://cds.team/taiga"),
  cache.id = FALSE, no.save = FALSE,
  taiga.api.version = getOption("default.taiga.api.version", 2),
  force.convert = F)
}
\arguments{
\item{datasetVersion.id}{DatasetVersion id}

\item{dataset.name}{Permaname of the dataset we want to retrieve data from}

\item{dataset.version}{Version of the dataset}

\item{transpose}{Will transpose all the matrices}

\item{cache.dir}{Path to the directory where to put your data in. !Not recommended to modify this}

\item{force.taiga}{Boolean to force the download of the data from Taiga instead of using the cache (if you modified the data for example)}

\item{taiga.url}{Url to taiga. !Not recommended to change this}

\item{cache.id}{Id of the cache used to store the file}

\item{no.save}{Boolean to not put the file in cache}

\item{taiga.api.version}{Version of the Taiga api to use. !Not receommended to change this}

\item{force.convert}{Boolean to force a new conversion of the data. Might be useful if the conversion has been interrupted}
}
\value{
Hash table of the form `filename`: `data`
}
\description{
Function to retrieve all the data available in a specific datasetVersion
}
