% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taigr.R
\name{load.all.from.taiga}
\alias{load.all.from.taiga}
\title{Load multiple datasets from taiga}
\usage{
load.all.from.taiga(info, ...)
}
\arguments{
\item{info}{named list of arguments to load.from.taiga}

\item{...}{extra arguments passed to load.from.taiga}
}
\value{
named list of datasets
}
\description{
Load multiple datasets from taiga
}
\examples{
datasets.info <- list(
    cnv = list(
        data.name = "ccle-copy-number-variants",
        data.version = 1),
    rpkm = list(
        data.name="ccle-rnaseq-gene-expression-rpkm-for-analysis-in-manuscripts-protein-coding-genes-only-hgnc-mapped",
        data.version = 3))
datasets <- load.all.from.taiga(datasets.info, transpose=TRUE)
}
