% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taigr.R
\name{load.from.taiga}
\alias{load.from.taiga}
\title{Load data from taiga}
\usage{
load.from.taiga(data.id = NULL, data.name = NULL,
  data.version = NULL, transpose = FALSE, data.dir = "~/.taiga",
  force.taiga = FALSE, taiga.url = getOption("default.taiga.url",
  "https://cds.team/taiga"), cache.id = FALSE, no.save = FALSE,
  quiet = FALSE,
  taiga.api.version = getOption("default.taiga.api.version", 2),
  force.convert = F, data.file = NULL)
}
\arguments{
\item{data.id}{The dataset ID in taiga.}

\item{data.name}{The dataset name in taiga.}

\item{data.version}{The dataset version number in taiga.}

\item{transpose}{transpose the data before returning it. the cached
version will not be transposed.}

\item{data.dir}{Where to look for and save cached version of the data.}

\item{force.taiga}{Force function to re-download data from taiga.}

\item{taiga.url}{Where is taiga?}

\item{cache.id}{use <id>.RData for filename instead of
<name>_<version>.RData}

\item{no.save}{Do not save dataset to cache.}

\item{quiet}{Do not print messages.}

\item{data.file}{file to load from within the dataset}
}
\value{
The dataset loaded into your R session.
}
\description{
Load data from taiga
}
