% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taigr.R
\name{pretty.print.taiga.info}
\alias{pretty.print.taiga.info}
\title{Prettily print taiga dataset info}
\usage{
pretting.print.taiga.info(info)
}
\arguments{
\item{info}{named list of arguments to load.from.taiga}
}
\description{
Prettily print taiga dataset info
}
