% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taigr.R
\docType{package}
\name{taigr}
\alias{taigr}
\alias{taigr-package}
\title{taigr.}
\description{
interface with taiga including a cache functionality
}
\details{
When loading data from taiga, it can look for a saved version
of the data object. If it doesn't find one, will load regularly from taiga
and write the object to disk.
}
