% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taigr.R
\name{visit.taiga.page}
\alias{visit.taiga.page}
\title{Visit Taiga page in web browser}
\usage{
visit.taiga.page(data.id = NULL, data.name = NULL,
  data.version = NULL, taiga.url = getOption("default.taiga.url",
  "https://cds.team/taiga"),
  taiga.api.version = getOption("default.taiga.api.version", 2))
}
\arguments{
\item{data.id}{The dataset ID in taiga.}

\item{data.name}{The dataset name in taiga.}

\item{data.version}{The dataset version number in taiga.}

\item{taiga.url}{Where is taiga?}
}
\value{
the full URL
}
\description{
Visit Taiga page in web browser
}
