% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taigr.R
\name{load.all.datafiles.from.taiga}
\alias{load.all.datafiles.from.taiga}
\title{Function to retrieve all the data available in a specific datasetVersion}
\usage{
load.all.datafiles.from.taiga(data.id = NULL, data.name = NULL,
  data.version = NULL, transpose = FALSE, data.dir = "~/.taiga",
  force.taiga = FALSE, taiga.url = getOption("default.taiga.url",
  "https://cds.team/taiga"), cache.id = FALSE, no.save = FALSE,
  taiga.api.version = getOption("default.taiga.api.version", 2),
  force.convert = F)
}
\arguments{
\item{data.id}{DatasetVersion id}

\item{data.name}{Permaname of the dataset we want to retrieve data from}

\item{data.version}{Version of the dataset}

\item{transpose}{Will transpose all the matrices}

\item{data.dir}{Path to the directory where to put your data in. !Not recommended to modify this}

\item{force.taiga}{Boolean to force the download of the data from Taiga instead of using the cache (if you modified the data for example)}

\item{taiga.url}{Url to taiga. !Not recommended to change this}

\item{no.save}{Boolean to not put the file in cache}

\item{taiga.api.version}{Version of the Taiga api to use. !Not receommended to change this}

\item{force.convert}{Boolean to force a new conversion of the data. Might be useful if the conversion has been interrupted}
}
\value{
Hash table of the form `filename`: `data`
}
\description{
Function to retrieve all the data available in a specific datasetVersion
}
