% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taigr.R
\name{load.from.taiga}
\alias{load.from.taiga}
\title{Load data from taiga}
\usage{
load.from.taiga(data.id = NULL, data.name = NULL, data.version = NULL,
  data.dir = "~/.taiga", force.taiga = FALSE,
  taiga.url = "http://datasci-dev:8999", no.save = FALSE, quiet = FALSE)
}
\arguments{
\item{data.id}{The dataset ID in taiga.}

\item{data.name}{The dataset name in taiga.}

\item{data.version}{The dataset version number in taiga.}

\item{data.dir}{Where to look for and save cached version of the data.}

\item{force.taiga}{Force function to re-download data from taiga.}

\item{taiga.url}{Where is taiga?}

\item{no.save}{Do not save dataset to cache.}

\item{quiet}{Do not print messages.}
}
\value{
The dataset loaded into your R session.
}
\description{
Load data from taiga
}

